Made service for SpringDetector and made SpringDetector delegate to it
