+++ dspace-api/src/main/java/org/dspace/statistics/factory/StatisticsServiceFactory.java
@@ -10,6 +10,8 @@ package org.dspace.statistics.factory;
 import org.dspace.services.factory.DSpaceServicesFactory;
 import org.dspace.statistics.service.ElasticSearchLoggerService;
 import org.dspace.statistics.service.SolrLoggerService;
+import org.dspace.statistics.util.SpiderDetector;
+import org.dspace.statistics.util.SpiderDetectorService;
 
 /**
  * Abstract factory to get services for the statistics package, use StatisticsServiceFactory.getInstance() to retrieve an implementation
@@ -21,6 +23,8 @@ public abstract class StatisticsServiceFactory {
     public abstract SolrLoggerService getSolrLoggerService();
 
     public abstract ElasticSearchLoggerService getElasticSearchLoggerService();
+
+    public abstract SpiderDetectorService getSpiderDetectorService();
 
     public static StatisticsServiceFactory getInstance()
     {
