+++ dspace-api/src/main/java/org/dspace/statistics/factory/StatisticsServiceFactoryImpl.java
@@ -10,6 +10,8 @@ package org.dspace.statistics.factory;
 import org.dspace.services.factory.DSpaceServicesFactory;
 import org.dspace.statistics.service.ElasticSearchLoggerService;
 import org.dspace.statistics.service.SolrLoggerService;
+import org.dspace.statistics.util.SpiderDetector;
+import org.dspace.statistics.util.SpiderDetectorService;
 
 /**
  * Factory implementation to get services for the statistics package, use StatisticsServiceFactory.getInstance() to retrieve an implementation
@@ -28,5 +30,10 @@ public class StatisticsServiceFactoryImpl extends StatisticsServiceFactory {
     public ElasticSearchLoggerService getElasticSearchLoggerService() {
         // In order to lazy load, we cannot autowire it and instead load it by name
         return DSpaceServicesFactory.getInstance().getServiceManager().getServiceByName("elasticSearchLoggerService", ElasticSearchLoggerService.class);
+    }
+
+    @Override
+    public SpiderDetectorService getSpiderDetectorService() {
+        return DSpaceServicesFactory.getInstance().getServiceManager().getServiceByName("spiderDetectorService", SpiderDetectorService.class);
     }
 }
