+++ dspace-api/src/main/java/org/dspace/statistics/util/SpiderDetector.java
@@ -7,62 +7,41 @@
  */
 package org.dspace.statistics.util;
 
-import java.io.BufferedReader;
 import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
-import java.util.Collections;
-import java.util.regex.Pattern;
 import javax.servlet.http.HttpServletRequest;
 
-import org.apache.commons.configuration.ConversionException;
-import org.apache.commons.lang.StringUtils;
-import org.dspace.services.ConfigurationService;
-import org.dspace.services.factory.DSpaceServicesFactory;
+import org.dspace.statistics.factory.StatisticsServiceFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * SpiderDetector is used to find IP's that are spiders...
- * In future someone may add Host Domains
- * to the detection criteria here.
+ * SpiderDetector delegates static methods to SpiderDetectorService, which is used to find IP's that are spiders...
  *
  * @author kevinvandevelde at atmire.com
  * @author ben at atmire.com
  * @author Mark Diggory (mdiggory at atmire.com)
+ * @author frederic at atmire.com
  */
 public class SpiderDetector {
 
     private static final Logger log = LoggerFactory.getLogger(SpiderDetector.class);
 
-    private Boolean useProxies;
-
-    private Boolean useCaseInsensitiveMatching;
+    //Service where all methods get delegated to, this is instantiated by a spring-bean defined in core-services.xml
+    private static SpiderDetectorService spiderDetectorService = StatisticsServiceFactory.getInstance().getSpiderDetectorService();
 
     /**
-     * Sparse HashTable structure to hold IP address ranges.
+     * Get an immutable Set representing all the Spider Addresses here
+     *
+     * @return a set of IP addresses as strings
      */
-    private IPTable table = null;
+    public static Set<String> getSpiderIpAddresses() {
 
-    /** Collection of regular expressions to match known spiders' agents. */
-    private final List<Pattern> agents
-            = Collections.synchronizedList(new ArrayList<Pattern>());
-
-    /** Collection of regular expressions to match known spiders' domain names. */
-    private final List<Pattern> domains
-            = Collections.synchronizedList(new ArrayList<Pattern>());
-
-    private ConfigurationService configurationService;
-
-    private static SpiderDetector spiderDetector = new SpiderDetector(DSpaceServicesFactory.getInstance().getConfigurationService());
-
-    public SpiderDetector(ConfigurationService configurationService) {
-        this.configurationService = configurationService;
+        spiderDetectorService.loadSpiderIpAddresses();
+        return spiderDetectorService.getTable().toSet();
     }
+
     /**
      * Utility method which reads lines from a file & returns them in a Set.
      *
@@ -73,134 +52,7 @@ public class SpiderDetector {
     public static Set<String> readPatterns(File patternFile)
             throws IOException
     {
-        Set<String> patterns = new HashSet<>();
-
-        if (!patternFile.exists() || !patternFile.isFile())
-        {
-            return patterns;
-        }
-
-        //Read our file & get all them patterns.
-        try (BufferedReader in = new BufferedReader(new FileReader(patternFile)))
-        {
-            String line;
-            while ((line = in.readLine()) != null) {
-                if (!line.startsWith("#")) {
-                    line = line.trim();
-
-                    if (!line.equals("")) {
-                        patterns.add(line);
-                    }
-                } else {
-                    //   ua.add(line.replaceFirst("#","").replaceFirst("UA","").trim());
-                    // ... add this functionality later
-                }
-            }
-        }
-        return patterns;
-    }
-
-    /**
-     * Get an immutable Set representing all the Spider Addresses here
-     *
-     * @return a set of IP addresses as strings
-     */
-    public static Set<String> getSpiderIpAddresses() {
-
-        spiderDetector.loadSpiderIpAddresses();
-        return spiderDetector.table.toSet();
-    }
-
-    /*
-     *  private loader to populate the table from files.
-     */
-    private synchronized void loadSpiderIpAddresses() {
-
-        if (table == null) {
-            table = new IPTable();
-
-            String filePath = configurationService.getProperty("dspace.dir");
-
-            try {
-                File spidersDir = new File(filePath, "config/spiders");
-
-                if (spidersDir.exists() && spidersDir.isDirectory()) {
-                    for (File file : spidersDir.listFiles()) {
-                        if (file.isFile())
-                        {
-                            for (String ip : readPatterns(file)) {
-                                log.debug("Loading {}", ip);
-                                if (!Character.isDigit(ip.charAt(0)))
-                                {
-                                    try {
-                                        ip = DnsLookup.forward(ip);
-                                        log.debug("Resolved to {}", ip);
-                                    } catch (IOException e) {
-                                        log.warn("Not loading {}:  {}", ip, e.getMessage());
-                                        continue;
-                                    }
-                                }
-                                table.add(ip);
-                            }
-                            log.info("Loaded Spider IP file: " + file);
-                        }
-                    }
-                } else {
-                    log.info("No spider file loaded");
-                }
-            }
-            catch (IOException | IPTable.IPFormatException e) {
-                log.error("Error Loading Spiders:" + e.getMessage(), e);
-            }
-
-        }
-
-    }
-
-    /**
-     * Load agent name patterns from all files in a single subdirectory of config/spiders.
-     *
-     * @param directory simple directory name (e.g. "agents").
-     *      "${dspace.dir}/config/spiders" will be prepended to yield the path to
-     *      the directory of pattern files.
-     * @param patternList patterns read from the files in {@code directory} will
-     *      be added to this List.
-     */
-    private void loadPatterns(String directory, List<Pattern> patternList)
-    {
-        String dspaceHome = configurationService.getProperty("dspace.dir");
-        File spidersDir = new File(dspaceHome, "config/spiders");
-        File patternsDir = new File(spidersDir, directory);
-        if (patternsDir.exists() && patternsDir.isDirectory())
-        {
-            for (File file : patternsDir.listFiles())
-            {
-                Set<String> patterns;
-                try
-                {
-                    patterns = readPatterns(file);
-                } catch (IOException ex)
-                {
-                    log.error("Patterns not read from {}:  {}",
-                            file.getPath(), ex.getMessage());
-                    continue;
-                }
-                //If case insensitive matching is enabled, lowercase the patterns so they can be lowercase matched
-                for (String pattern : patterns) {
-                    if(isUseCaseInsensitiveMatching()) {
-                        pattern = StringUtils.lowerCase(pattern);
-                    }
-                    patternList.add(Pattern.compile(pattern));
-                }
-
-
-                log.info("Loaded pattern file:  {}", file.getPath());
-            }
-        }
-        else
-        {
-            log.info("No patterns loaded from {}", patternsDir.getPath());
-        }
+        return spiderDetectorService.readPatterns(patternFile);
     }
 
     /**
@@ -215,84 +67,11 @@ public class SpiderDetector {
      * @param agent User-Agent header value, or null.
      * @return true if the client matches any spider characteristics list.
      */
-    public boolean isSpiderAgent(String clientIP, String proxyIPs,
-                                   String hostname, String agent)
-    {
-        // See if any agent patterns match
-        if (null != agent)
-        {
-            synchronized(agents)
-            {
-                if (agents.isEmpty())
-                    loadPatterns("agents", agents);
-            }
-
-            if(isUseCaseInsensitiveMatching()) {
-                agent = StringUtils.lowerCase(agent);
-                hostname = StringUtils.lowerCase(hostname);
-            }
-
-            for (Pattern candidate : agents) {
-
-                // prevent matcher() invocation from a null Pattern object
-                if (null != candidate && candidate.matcher(agent).find()) {
-                    return true;
-                }
-
-
-            }
-        }
-
-        // No.  See if any IP addresses match
-        if (isUseProxies() && proxyIPs != null) {
-            /* This header is a comma delimited list */
-            for (String xfip : proxyIPs.split(",")) {
-                if (isSpiderAgent(xfip))
-                {
-                    return true;
-                }
-            }
-        }
-
-        if (isSpiderAgent(clientIP))
-            return true;
-
-        // No.  See if any DNS names match
-        if (null != hostname)
-        {
-            synchronized(domains)
-            {
-                if (domains.isEmpty())
-                    loadPatterns("domains", domains);
-            }
-            for (Pattern candidate : domains)
-            {
-                // prevent matcher() invocation from a null Pattern object
-                if (null != candidate && candidate.matcher(hostname).find())
-                {
-                    return true;
-                }
-            }
-        }
-
-        // Not a known spider.
-        return false;
-    }
-
     public static boolean isSpider(String clientIP, String proxyIPs,
                                    String hostname, String agent)
     {
-        return spiderDetector.isSpiderAgent(clientIP, proxyIPs, hostname, agent);
+        return spiderDetectorService.isSpider(clientIP, proxyIPs, hostname, agent);
     }
-
-    public boolean isSpiderAgent(HttpServletRequest request)
-    {
-        return isSpiderAgent(request.getRemoteAddr(),
-                request.getHeader("X-Forwarded-For"),
-                request.getRemoteHost(),
-                request.getHeader("User-Agent"));
-    }
-
 
     /**
      * Static Service Method for testing spiders against existing spider files.
@@ -302,26 +81,7 @@ public class SpiderDetector {
      */
     public static boolean isSpider(HttpServletRequest request)
     {
-        return spiderDetector.isSpiderAgent(request);
-    }
-
-    public boolean isSpiderAgent(String ip) {
-
-        if (table == null) {
-            loadSpiderIpAddresses();
-        }
-
-        try {
-            if (table.contains(ip)) {
-                return true;
-            }
-        } catch (Exception e) {
-            return false;
-        }
-
-        return false;
-
-
+        return spiderDetectorService.isSpider(request);
     }
 
     /**
@@ -331,31 +91,7 @@ public class SpiderDetector {
      * @return if is spider IP
      */
     public static boolean isSpider(String ip) {
-
-        return spiderDetector.isSpiderAgent(ip);
-
-
-    }
-
-    private boolean isUseCaseInsensitiveMatching() {
-        if (useCaseInsensitiveMatching == null) {
-            try {
-                useCaseInsensitiveMatching = configurationService.getBooleanProperty("usage-statistics.bots.case-insensitive");
-            } catch (ConversionException e) {
-                useCaseInsensitiveMatching = false;
-                log.warn("Please use a boolean value for usage-statistics.bots.case-insensitive");
-            }
-        }
-
-        return useCaseInsensitiveMatching;
-    }
-
-    private boolean isUseProxies() {
-        if(useProxies == null) {
-            useProxies = configurationService.getBooleanProperty("useProxies");
-        }
-
-        return useProxies;
+        return spiderDetectorService.isSpider(ip);
     }
 
 }
